Drop XML in favor of one module = one menu paradigm
