+++ pooltool/ani/modes/cam_load.py
@@ -3,10 +3,10 @@
 from direct.gui.DirectGui import DGG
 
 import pooltool.ani.tasks as tasks
+import pooltool.ani.utils as autils
 from pooltool.ani.action import Action
 from pooltool.ani.camera import cam
 from pooltool.ani.globals import Global
-from pooltool.ani.menu import TextOverlay
 from pooltool.ani.modes.datatypes import BaseMode, Mode
 from pooltool.ani.mouse import MouseMode, mouse
 
@@ -45,7 +45,7 @@ class CamLoadMode(BaseMode):
         self.cam_load_slots.hide()
 
     def render_camera_load_buttons(self):
-        self.cam_load_slots = TextOverlay(
+        self.cam_load_slots = autils.TextOverlay(
             title="Release key with moused hovered over desired save slot",
             frame_color=(0, 0, 0, 0.2),
             title_pos=(0, 0, 0.45),
