+++ pooltool/ani/modes/cam_save.py
@@ -3,10 +3,10 @@
 from direct.gui.DirectGui import DGG
 
 import pooltool.ani.tasks as tasks
+import pooltool.ani.utils as autils
 from pooltool.ani.action import Action
 from pooltool.ani.camera import cam
 from pooltool.ani.globals import Global
-from pooltool.ani.menu import TextOverlay
 from pooltool.ani.modes.datatypes import BaseMode, Mode
 from pooltool.ani.mouse import MouseMode, mouse
 
@@ -42,7 +42,7 @@ class CamSaveMode(BaseMode):
         del self.selection
 
     def render_camera_save_buttons(self):
-        self.cam_save_slots = TextOverlay(
+        self.cam_save_slots = autils.TextOverlay(
             title="Release key with moused hovered over desired save slot",
             frame_color=(0, 0, 0, 0.2),
             title_pos=(0, 0, 0.45),
