--- pooltool/config/menus/game_setup.xml
@@ -1,17 +1,0 @@
-<menu name="game_setup">
-    <title>Game Setup</title>
-    <backbutton>
-        <func>func_go_main_menu</func>
-    </backbutton>
-    <button>
-        <name>Play now</name>
-        <func>func_play_now</func>
-        <description>Enter the game with the currently selected settings</description>
-    </button>
-    <text>
-        There used to be some customizable options here, but due to some dramatic
-        changes to the codebase, the archaic menu system no longer plugs into the
-        backend. That will change one day, but today is not that day :( For now, just
-        click "Play now"
-    </text>
-</menu>
