--- pooltool/config/menus/main_menu.xml
@@ -1,18 +1,0 @@
-<menu name="main_menu">
-    <title>Main Menu</title>
-    <button>
-        <name>New Game</name>
-        <func>func_go_game_setup</func>
-        <description>Play some pool. Shortcut: n</description>
-    </button>
-    <button>
-        <name>Settings</name>
-        <func>func_go_settings</func>
-        <description>View and edit available settings.</description>
-    </button>
-    <button>
-        <name>Quit</name>
-        <func>func_quit_pooltool</func>
-        <description>Closes pooltool. Shortcut: esc</description>
-    </button>
-</menu>
